+++ docs/_jinja_funcs.py
@@ -1,6 +1,6 @@
 """Example jinja filters/tests for the documentation."""
 
 
-def is_big(value):
+def is_big(value: int) -> bool:
     """Test if a value is bigger than 100."""
     return value > 100